BLE: - Move Gadgetbridge related code into separate source file
