+++ app/src/ble/CMakeLists.txt
@@ -1,6 +1,7 @@
 target_sources(app PRIVATE ble_aoa.c)
 target_sources(app PRIVATE ble_comm.c)
 target_sources(app PRIVATE ble_transport.c)
+target_sources(app PRIVATE gadgetbridge/ble_gadgetbridge.c)
 target_sources_ifdef(CONFIG_BLE_USES_IOS app PRIVATE ble_ams.c)
 target_sources_ifdef(CONFIG_BLE_USES_IOS app PRIVATE ble_ancs.c)
 target_sources_ifdef(CONFIG_BLE_USES_IOS app PRIVATE ble_cts.c)
